move writers/writers.hpp -> writers.hpp
